WPF Layout port in progress.

+++ Source/Alternet.UI/Controls/Grid/GridLengthConverter.cs
@@ -1,10 +1,14 @@
 #nullable disable
 
-using System;
 using System.ComponentModel;
+using System.Windows;
+using System;
+using System.Security;
 using System.ComponentModel.Design.Serialization;
+using System.Diagnostics;
 using System.Globalization;
 using System.Reflection;
+using System.Windows.Markup;
 
 namespace Alternet.UI
 {
@@ -12,7 +16,7 @@ namespace Alternet.UI
     /// GridLengthConverter - Converter class for converting 
     /// instances of other types to and from GridLength instances.
     /// </summary> 
-    public class GridLengthConverter : TypeConverter
+    public class GridLengthConverter: TypeConverter
     {
         //-------------------------------------------------------------------
         //
@@ -33,7 +37,7 @@ namespace Alternet.UI
         /// <c>false</c> otherwise.
         /// </returns>
         public override bool CanConvertFrom(
-            ITypeDescriptorContext typeDescriptorContext,
+            ITypeDescriptorContext typeDescriptorContext, 
             Type sourceType)
         {
             // We can only handle strings, integral and floating types
@@ -51,7 +55,7 @@ namespace Alternet.UI
                 case TypeCode.UInt32:
                 case TypeCode.UInt64:
                     return true;
-                default:
+                default: 
                     return false;
             }
         }
@@ -67,11 +71,11 @@ namespace Alternet.UI
         /// <c>false</c> otherwise.
         /// </returns>
         public override bool CanConvertTo(
-            ITypeDescriptorContext typeDescriptorContext,
-            Type destinationType)
+            ITypeDescriptorContext typeDescriptorContext, 
+            Type destinationType) 
         {
-            return (destinationType == typeof(InstanceDescriptor)
-                    || destinationType == typeof(string));
+            return (    destinationType == typeof(InstanceDescriptor) 
+                    ||  destinationType == typeof(string)   );
         }
 
         /// <summary>
@@ -91,8 +95,8 @@ namespace Alternet.UI
         /// and is not a valid type which can be converted to a GridLength.
         /// </exception>
         public override object ConvertFrom(
-            ITypeDescriptorContext typeDescriptorContext,
-            CultureInfo cultureInfo,
+            ITypeDescriptorContext typeDescriptorContext, 
+            CultureInfo cultureInfo, 
             object source)
         {
             if (source != null)
@@ -107,12 +111,12 @@ namespace Alternet.UI
                     double value;
                     GridUnitType type;
 
-                    value = Convert.ToSingle(source, cultureInfo);
+                    value = Convert.ToDouble(source, cultureInfo);
 
                     if (DoubleUtil.IsNaN(value))
                     {
                         //  this allows for conversion from Width / Height = "Auto" 
-                        value = 1.0f;
+                        value = 1.0;
                         type = GridUnitType.Auto;
                     }
                     else
@@ -144,7 +148,7 @@ namespace Alternet.UI
         /// or if the destinationType isn't one of the valid destination types.
         /// </exception>
         public override object ConvertTo(
-            ITypeDescriptorContext typeDescriptorContext,
+            ITypeDescriptorContext typeDescriptorContext, 
             CultureInfo cultureInfo,
             object value,
             Type destinationType)
@@ -154,14 +158,14 @@ namespace Alternet.UI
                 throw new ArgumentNullException("destinationType");
             }
 
-            if (value != null
-                && value is GridLength)
+            if (    value != null
+                &&  value is GridLength )
             {
                 GridLength gl = (GridLength)value;
 
-                if (destinationType == typeof(string))
-                {
-                    return (ToString(gl, cultureInfo));
+                if (destinationType == typeof(string)) 
+                { 
+                    return (ToString(gl, cultureInfo)); 
                 }
 
                 if (destinationType == typeof(InstanceDescriptor))
